test: Move nocks to their own folder
